+++ provider/datagen/src/transform/cldr/list/mod.rs
@@ -145,79 +145,3 @@ macro_rules! implement {
 implement!(AndListV1Marker);
 implement!(OrListV1Marker);
 implement!(UnitListV1Marker);
-
-#[cfg(test)]
-mod tests {
-    use icu_list::{ListFormatter, ListLength};
-    use icu_locid::locale;
-    use writeable::assert_writeable_eq;
-
-    macro_rules! test {
-        ($locale:literal, $type:ident, $(($input:expr, $output:literal),)+) => {
-            let f = ListFormatter::$type(
-                &crate::DatagenProvider::for_test(),
-                &locale!($locale).into(),
-                ListLength::Wide
-            ).unwrap();
-            $(
-                assert_writeable_eq!(f.format($input.iter()), $output);
-            )+
-        };
-    }
-
-    #[test]
-    fn test_basic() {
-        test!(
-            "fr",
-            try_new_or_with_length_unstable,
-            (["A", "B"], "A ou B"),
-        );
-    }
-
-    #[test]
-    fn test_spanish() {
-        test!(
-            "es",
-            try_new_and_with_length_unstable,
-            (["x", "Mallorca"], "x y Mallorca"),
-            (["x", "Ibiza"], "x e Ibiza"),
-            (["x", "Hidalgo"], "x e Hidalgo"),
-            (["x", "Hierva"], "x y Hierva"),
-        );
-
-        test!(
-            "es",
-            try_new_or_with_length_unstable,
-            (["x", "Ibiza"], "x o Ibiza"),
-            (["x", "Okinawa"], "x u Okinawa"),
-            (["x", "8 más"], "x u 8 más"),
-            (["x", "8"], "x u 8"),
-            (["x", "87 más"], "x u 87 más"),
-            (["x", "87"], "x u 87"),
-            (["x", "11 más"], "x u 11 más"),
-            (["x", "11"], "x u 11"),
-            (["x", "110 más"], "x o 110 más"),
-            (["x", "110"], "x o 110"),
-            (["x", "11.000 más"], "x u 11.000 más"),
-            (["x", "11.000"], "x u 11.000"),
-            (["x", "11.000,92 más"], "x u 11.000,92 más"),
-            (["x", "11.000,92"], "x u 11.000,92"),
-        );
-
-        test!(
-            "es-AR",
-            try_new_and_with_length_unstable,
-            (["x", "Ibiza"], "x e Ibiza"),
-        );
-    }
-
-    #[test]
-    fn test_hebrew() {
-        test!(
-            "he",
-            try_new_and_with_length_unstable,
-            (["x", "יפו"], "x ויפו"),
-            (["x", "Ibiza"], "x ו-Ibiza"),
-        );
-    }
-}
