--- provider/datagen/tests/data/cldr/cldr-misc-full/main/he/listPatterns.json
@@ -1,68 +1,0 @@
-{
-  "main": {
-    "he": {
-      "identity": {
-        "version": {
-          "_cldrVersion": "43"
-        },
-        "language": "he"
-      },
-      "listPatterns": {
-        "listPattern-type-standard": {
-          "start": "{0}, {1}",
-          "middle": "{0}, {1}",
-          "end": "{0} ו{1}",
-          "2": "{0} ו{1}"
-        },
-        "listPattern-type-or": {
-          "start": "{0}, {1}",
-          "middle": "{0}, {1}",
-          "end": "{0} או {1}",
-          "2": "{0} או {1}"
-        },
-        "listPattern-type-or-narrow": {
-          "start": "{0}, {1}",
-          "middle": "{0}, {1}",
-          "end": "{0} או {1}",
-          "2": "{0} או {1}"
-        },
-        "listPattern-type-or-short": {
-          "start": "{0}, {1}",
-          "middle": "{0}, {1}",
-          "end": "{0} או {1}",
-          "2": "{0} או {1}"
-        },
-        "listPattern-type-standard-narrow": {
-          "start": "{0}, {1}",
-          "middle": "{0}, {1}",
-          "end": "{0} ו{1}",
-          "2": "{0} ו{1}"
-        },
-        "listPattern-type-standard-short": {
-          "start": "{0}, {1}",
-          "middle": "{0}, {1}",
-          "end": "{0} ו{1}",
-          "2": "{0} ו{1}"
-        },
-        "listPattern-type-unit": {
-          "start": "{0}, {1}",
-          "middle": "{0}, {1}",
-          "end": "{0} ו-{1}",
-          "2": "{0}, {1}"
-        },
-        "listPattern-type-unit-narrow": {
-          "start": "{0} {1}",
-          "middle": "{0} {1}",
-          "end": "{0} {1}",
-          "2": "{0} {1}"
-        },
-        "listPattern-type-unit-short": {
-          "start": "{0}, {1}",
-          "middle": "{0}, {1}",
-          "end": "{0}, {1}",
-          "2": "{0}, {1}"
-        }
-      }
-    }
-  }
-}
