+++ tools/testdata-scripts/globs.rs.data
@@ -38,7 +38,6 @@ const CLDR_JSON_GLOB: &[&str] = &[
     "cldr-localenames-full/main/$LOCALES/variants.json",
     "cldr-misc-full/main/$LOCALES/characters.json",
     "cldr-misc-full/main/$LOCALES/listPatterns.json",
-    "cldr-misc-full/main/he/listPatterns.json", // required for list transformer test
     "cldr-numbers-full/main/$LOCALES/numbers.json",
     "cldr-person-names-full/main/$LOCALES/personNames.json",
 ];
